third_party/WebKit/LayoutTests/typedcssom/cssPositionValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(() => {
  assert_equals(
      new CSSPositionValue(CSS.px(50), CSS.em(2)).toString(), '50px 2em');
}, "toString returns the x and y cssText separated by a space");

</script>
